avoid javascript errors due to invisible control
Using display: none; or visbility: hidden; will throw an error when the
user tries to submit the form without selecting a file (as the browser
fails to show the "required" tooltip).

In the case of a required file input, the form would just not submit and
the user wouldn't understand the problem. We have to lift the required
attribute from the input to avoid this behavior. Server side validation
is not affected by this.